The drawer follows the finger: the stored fraction is written on release only

+++ Packages/YomidoriCore/Sources/YomidoriKit/Capture/CaptureDrawer.swift
@@ -8,7 +8,7 @@ struct CaptureDrawer<Content: View, Buttons: View>: View {
     let hasStill: Bool
     let screenHeight: CGFloat
     @Binding var fraction: Double
-    let settled: () -> Void
+    let settled: (Double) -> Void
     @ViewBuilder var content: () -> Content
     @ViewBuilder var buttons: () -> Buttons
